Add files via upload
Version 1 of pseudo_risk, a bare bones tool used to describe the risk associated with a particular move from a starting board state to an ending board state, assuming that both board states are legal and that the end board state is achievable. The pseudo prefix implies that a more full bodied and robust risk tool is coming down the pipeline...eventually.